+++ libmisc/yesno.c
@@ -70,7 +70,7 @@ bool yes_or_no (bool read_only)
 	 * Get a line and see what the first character is.
 	 */
 	/* TODO: use gettext */
-	if (fgets (buf, sizeof buf, stdin) == buf) {
+	if (fgets (buf, (int) sizeof buf, stdin) == buf) {
 		return buf[0] == 'y' || buf[0] == 'Y';
 	}
